requirements.txt
clip==1.0
numpy==1.24.2
openai==0.27.2
Pillow==9.4.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
streamlit==1.20.0
torch==2.0.0
